services/model-manager/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kitty-model-manager"
version = "0.1.0"
description = "Terminal TUI model manager for KITTY's llama.cpp servers"
authors = [{name = "KITTY Team"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["llama.cpp", "model-management", "tui", "terminal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "textual>=0.47.0",      # Modern TUI framework
    "rich>=13.7.0",         # Terminal styling (already used in CLI)
    "pydantic>=2.5.0",      # Data validation
    "pydantic-settings>=2.0.0",  # Settings management
    "httpx>=0.27.0",        # HTTP client (already used)
    "python-dotenv>=1.0.0", # .env parsing (already used)
    "psutil>=5.9.0",        # Process/system monitoring
    "pyyaml>=6.0.0",        # YAML config files
    "typer[all]>=0.12.3",   # CLI framework (already used)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
kitty-model-manager = "model_manager.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
model_manager = ["py.typed"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "D",  # pydocstyle
    "UP", # pyupgrade
    "B",  # flake8-bugbear
]
ignore = [
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D203", # 1 blank line required before class docstring
    "D213", # Multi-line docstring summary should start at the second line
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]  # No docstrings required in tests

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "textual.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "-v",
]
markers = [
    "unit: Unit tests that don't require external dependencies",
    "integration: Integration tests requiring llama.cpp server",
]
